Finalize Tests for ENSRegistry & ENSRegistryWithFallback & Delete ENSRegistryWithFallback
ENS discovered an exploit, which required them to migrate away from their original contracts.
ENSRegistryWithFallback was deployed as a means to migrate smoothly from the old ENSRegistry
the new ENSRegistry. Since we will simple deploy the updated ENSRegistry, we only need to
test this newer contract with the expanded tests from ENSRegistryWithFallback, and not use
the migrate-only ENSRegistryWithFallback contract at all.

'virtual override' was also removed from all functions, because it is not needed, as there
will no longer be a contract inheriting from ENSRegistry (formerly ENSRegistryWithFallback)
that would need to override any of the ENS Interface Functions a second time.